Make all changes in test_rwx.py
