+++ manager/integration/tests/test_rwx.py
@@ -15,12 +15,20 @@ from common import DEFAULT_STATEFULSET_TIMEOUT, DEFAULT_STATEFULSET_INTERVAL
 from common import wait_delete_pod, wait_for_pod_remount
 from common import get_core_api_client, write_pod_volume_random_data
 from common import create_pvc_spec, make_deployment_with_pvc  # NOQA
-from common import core_api, statefulset, pvc, pod, client, nfs  # NOQA
+from common import core_api, statefulset, pvc, pod, client  # NOQA
 from common import RETRY_COUNTS, RETRY_INTERVAL
 from backupstore import set_random_backupstore # NOQA
 from multiprocessing import Pool
 
 import time
+import pytest
+import subprocess
+
+LONGHORN_NFS_INSTALLATION_URL = \
+    "https://raw.githubusercontent.com/longhorn/" \
+    "longhorn/master/deploy/prerequisite/longhorn" \
+    "-nfs-installation.yaml"
+LONGHORN_NFS_DAEMONSET_NAME = "longhorn-nfs-installation"
 
 
 def write_data_into_pod(pod_name_and_data_path):
@@ -28,6 +36,46 @@ def write_data_into_pod(pod_name_and_data_path):
     core_api = get_core_api_client()  # NOQA
     write_pod_volume_random_data(core_api, pod_info[0], pod_info[1],
                                  DATA_SIZE_IN_MB_3)
+
+
+@pytest.fixture(scope="module", autouse="True")
+def nfs(request):
+
+    api = get_core_api_client()
+    apps_api = get_apps_api_client()
+
+    cmd = ["kubectl", "apply", "-f", LONGHORN_NFS_INSTALLATION_URL]
+    subprocess.check_output(cmd)
+
+    node_list = api.list_node()
+    cmd = ["kubectl get pod -l app={} | "
+           "awk 'NR>1 {{print $1}}'".format(LONGHORN_NFS_DAEMONSET_NAME)]
+    for i in range(RETRY_COUNTS):
+        nfs_pods = {}
+        pod_list = \
+            subprocess.check_output(cmd, shell=True).decode("utf-8").split()
+
+        for pod_name in pod_list:
+            nfs_pods[pod_name] = \
+                api.read_namespaced_pod(name=pod_name,
+                                        namespace='default').status.phase
+
+        if len(nfs_pods) == len(node_list.items):
+            if all(value == "Running" for value in nfs_pods.values()):
+                break
+
+        time.sleep(RETRY_INTERVAL)
+
+    if not \
+            all(value == "Running" for value in nfs_pods.values()) \
+            or not len(nfs_pods) == len(node_list.items):
+        raise Exception("Longhorn-nfs not installed on all nodes")
+
+    def finalizer():
+        apps_api.delete_namespaced_daemon_set(name=LONGHORN_NFS_DAEMONSET_NAME,
+                                              namespace='default')
+
+    request.addfinalizer(finalizer)
 
 
 def test_rwx_with_statefulset_multi_pods(core_api, statefulset):  # NOQA
